team-17/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
jinja2==3.1.4
beautifulsoup4==4.12.0


# AI/ML
google-generativeai==0.8.3
chromadb==0.5.23
numpy<2.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Data Processing
pydantic==2.10.3
pydantic-settings==2.7.0

# Utilities
python-dotenv==1.0.1
mistune==3.0.2
beautifulsoup4>=4.12,<5.0

# logging fix
posthog>=2.4.0,<6.0.0

# Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0
